pyproject.toml
[project]
name = "eval3-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pypdf>=6.5.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "qdrant-client[fastembed]>=1.16.2",
    "supabase>=2.27.0",
]
